Renamed Carousel component to Gallery for synergy with business doc
