+++ client/src/reducers/galleryReducer.js
